refactor: /sankey-svg page.tsx から型・定数・純粋関数を抽出
1646行のモノリシックファイルから475行を3つのモジュールに分離:
- types/sankey-svg.ts: 型定義 (RawNode, RawEdge, GraphData, LayoutNode, LayoutLink)
- app/lib/sankey-svg-constants.ts: 定数 + 純粋関数 (COL_MAP, MARGIN, getNodeColor, ribbonPath, formatYen 等)
- app/lib/sankey-svg-filter.ts: フィルタ・レイアウト (filterTopN, computeLayout)

Closes #116
